refactor(library/math): extracts `numberTaxonomy` into dedicated file
