+++ src/library/math/operator/predicate/isOperable.test.ts
@@ -5,30 +5,12 @@ import {
 } from 'vitest';
 
 import {
+  numberTaxonomy,
+} from '../../numberTaxonomy';
+
+import {
   isOperable,
 } from './isOperable';
-
-const numberTaxonomy = {
-  signed: {
-    negative: /**/-1,
-    positive: /**/+1,
-  },
-  fractional: {
-    rational      : 1 / 2,
-    irrational    : Math.PI,
-    transcendental: Math.E,
-  },
-  infinite: {
-    positive: Infinity,
-    negative: -Infinity,
-  },
-  runtime: {
-    min: Number.MIN_VALUE,
-    max: Number.MAX_VALUE,
-  },
-  origin   : 0,
-  undefined: NaN,
-} as const;
 
 describe(isOperable, () => {
   const theSoleInoperableNumber = numberTaxonomy.undefined;
